Creates a function for starting consumers
