--- backend/rabbitmq/startConsumers.js
@@ -1,7 +1,0 @@
-const { startTweetConsumer } = require("./consumers/tweets");
-
-const startConsumers = async (connection) => {
-  await startTweetConsumer(connection);
-};
-
-module.exports = startConsumers;
